consolidate deposit/withdraw flow into transferdialogs
