+++ src/views/dialogs/TransferDialogs/DepositDialog2/TokenSelect.tsx
